+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 ## 0.4.6
 
 - Added on_mouse_button callback (#184, #185)
+- Added on_tick callback (#187)
 
 ## 0.4.5